Decreased stack space of IdentifiedPeptidoformData by 25%

+++ rustyms/src/identification/formats/basic_csv.rs
@@ -146,7 +146,7 @@ impl MetaData for BasicCSVData {
         None
     }
 
-    fn protein_location(&self) -> Option<Range<usize>> {
+    fn protein_location(&self) -> Option<Range<u16>> {
         None
     }
 }
