+++ rustyms/src/identification/formats/deepnovofamily.rs
@@ -232,7 +232,7 @@ impl MetaData for DeepNovoFamilyData {
         None
     }
 
-    fn protein_location(&self) -> Option<Range<usize>> {
+    fn protein_location(&self) -> Option<Range<u16>> {
         None
     }
 }
