+++ rustyms/src/identification/formats/fasta.rs
@@ -825,7 +825,7 @@ impl MetaData for FastaData {
         None
     }
 
-    fn protein_location(&self) -> Option<Range<usize>> {
+    fn protein_location(&self) -> Option<Range<u16>> {
         None
     }
 }
