+++ rustyms/src/identification/formats/pepnet.rs
@@ -149,7 +149,7 @@ impl MetaData for PepNetData {
         None
     }
 
-    fn protein_location(&self) -> Option<Range<usize>> {
+    fn protein_location(&self) -> Option<Range<u16>> {
         None
     }
 }
